# backend/swagger.yml
openapi: 3.0.0
info:
  title: Backend API
  description: This is the API documentation for the backend used in the Automation QA Interview task.
  version: 1.0.0

servers:
  - url: http://localhost:8000
    description: Local server

paths:
  /signup/:
    post:
      summary: Sign up a new user
      requestBody:
        description: User credentials for signing up
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '200':
          description: User successfully signed up
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string
        '400':
          description: Invalid input

  /login/:
    post:
      summary: Log in an existing user
      requestBody:
        description: User credentials for logging in
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '200':
          description: User successfully logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
        '400':
          description: Incorrect username or password

  /profile/:
    get:
      summary: Get user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                  name:
                    type: string
                  last_name:
                    type: string
        '401':
          description: Unauthorized

    put:
      summary: Update user profile
      requestBody:
        description: Only name and last name can be updated
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                last_name:
                  type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  name:
                    type: string
                  last_name:
                    type: string
        '401':
          description: Unauthorized

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT